+++ src/test/java/org/folio/mr/api/BaseIT.java
@@ -115,7 +115,8 @@ public class BaseIT {
 
   @Container
   private static final KafkaContainer kafka = new KafkaContainer(
-    DockerImageName.parse("apache/kafka-native:3.8.0"));
+    DockerImageName.parse("apache/kafka-native:3.8.0"))
+    .withStartupAttempts(3);
 
   @DynamicPropertySource
   static void overrideProperties(DynamicPropertyRegistry registry) {
@@ -133,7 +134,7 @@ public class BaseIT {
   }
 
   @AfterEach
-  public void afterEachTest() {
+  void afterEachTest() {
     contextSetter.close();
   }